Se separo por archivo los modelos, al igual los modelos de Dto se movieron a una carpeta llamada Model en data.
