+++ app/src/main/java/com/varqulabs/dolarblue/calculator/data/remote/DolarBlueService.kt
@@ -1,5 +1,6 @@
 package com.varqulabs.dolarblue.calculator.data.remote
 
+import com.varqulabs.dolarblue.calculator.data.remote.model.DolarBlueDto
 import retrofit2.http.GET
 
 interface DolarBlueService {
