+++ app/src/main/java/com/varqulabs/dolarblue/calculator/data/remote/model/DolarBlueDto.kt
@@ -1,0 +1,18 @@
+package com.varqulabs.dolarblue.calculator.data.remote.model
+
+import com.squareup.moshi.Json
+import com.varqulabs.dolarblue.calculator.domain.model.BlueValue
+import com.varqulabs.dolarblue.calculator.domain.model.DolarBlue
+
+data class DolarBlueDto(
+    @Json(name = "blue")
+    val blue: BlueValueDto? = BlueValueDto(),
+    @Json(name = "last_update")
+    val last_update: String? = "",
+)
+
+fun DolarBlueDto.mapToEntity() = DolarBlue(
+    blue = blue?.mapToEntity() ?: BlueValue(),
+    lastUpdate = last_update.orEmpty(),
+)
+
