+++ app/src/main/java/com/varqulabs/dolarblue/calculator/domain/model/DolarBlue.kt
@@ -1,0 +1,7 @@
+package com.varqulabs.dolarblue.calculator.domain.model
+
+
+data class DolarBlue(
+    val blue: BlueValue = BlueValue(),
+    val lastUpdate: String = "",
+)
