+++ app/src/main/java/com/varqulabs/dolarblue/calculator/domain/repository/DolarBlueRepository.kt
@@ -1,9 +1,9 @@
 package com.varqulabs.dolarblue.calculator.domain.repository
 
-import com.varqulabs.dolarblue.calculator.domain.model.DolarBlueEntity
+import com.varqulabs.dolarblue.calculator.domain.model.DolarBlue
 
 interface DolarBlueRepository {
 
-    suspend fun getDolarBlue(): DolarBlueEntity
+    suspend fun getDolarBlue(): DolarBlue
 
 }
